YLists.Categorization/YLists.Categorization.Server/requirements.txt
fastparquet==0.8.1
grpcio==1.46.3
grpcio-tools==1.46.3
grpcio-reflection==1.46.3
nltk==3.7
numpy==1.22.4
pandas==1.4.2
pymorphy2==0.9.1
scikit-learn==1.1.1
